[core] fix the issue with excessive memory consumption during strategy optimization.
It seems that passing the cerebro instance as callback to the Pool.imap will not always correctly garbage collect the serialized cerebro instance.
The fix will use a regular global function as a callback and pass the cerebro instance as part of the iterable collection of parameters.
This way the serialized cerebro instance could be forcefully garbage collected upon callback completion.
This results in near constant memory consumption during the optimization process.